refactor(tests): hoist duplicated in-memory WFL drivers into tests/common
Three near-identical "run WFL source in memory" driver families were copy-pasted
across 12 integration test files, while `tests/common/mod.rs` held exactly one
helper. Each copy re-implemented the same lex -> parse -> interpret pipeline plus
its own `get_global`/`expect_*` accessors.

Hoisted into `tests/common/mod.rs`:

  run_wfl        -> Result<Interpreter, String>   (5 files)
  run_wfl_code   -> Result<Value, String>         (3 files)
  run_wfl_ok     -> Interpreter, panicking        (3 files)
  get_global / get_var / get_text / get_number / expect_* accessors

  12 files changed, 150 insertions(+), 379 deletions(-)   net -229 lines

`tests/http_connection_reuse_retry_test.rs` was deliberately left untouched: its
`get_text` matches on `Option<Value>` directly instead of routing through a
panicking `get_var`, so the missing-variable path panics differently. That is a
real behavioral divergence, not a cosmetic one, so its helpers stay local. A
sweep for the same direct-match shape found it only in that file plus
`overload_test.rs` and `static_container_member_test.rs`, none of which are
touched here.

No test function, assertion, input, or attribute was modified -- only helper
definitions and imports. Cosmetic `map_err` wording was unified after confirming
no test asserts on those strings.

Coverage evidence: `--list` test-name inventories were captured for all 12
targets before and after and diffed line by line -- every target's set of test
names is byte-identical, and the full-suite count is unchanged at 2096 passed.

Gates: cargo test --all 2096 passed / 0 failed / 27 ignored (159 suites); fmt
clean; clippy --all-targets --all-features -D warnings clean;
run_integration_tests.sh 136 passed / 0 failed / 24 skipped.
