+++ tests/filesystem_mode_test.rs
@@ -9,38 +9,10 @@
 // unsupported error — never a silent no-op, which is the failure mode the issue
 // complains about.
 
-use wfl::interpreter::Interpreter;
 use wfl::interpreter::value::Value;
-use wfl::lexer::lex_wfl_with_positions;
-use wfl::parser::Parser;
 
-async fn run_wfl(code: &str) -> Result<Interpreter, String> {
-    let tokens = lex_wfl_with_positions(code);
-    let mut parser = Parser::new(&tokens);
-    let ast = parser.parse().map_err(|e| format!("Parse error: {e:?}"))?;
-
-    let mut interpreter = Interpreter::new();
-    interpreter
-        .interpret(&ast)
-        .await
-        .map_err(|e| format!("Runtime error: {e:?}"))?;
-    Ok(interpreter)
-}
-
-fn get_global(interpreter: &Interpreter, name: &str) -> Value {
-    interpreter
-        .global_env()
-        .borrow()
-        .get(name)
-        .unwrap_or_else(|| panic!("Variable '{name}' not found"))
-}
-
-fn expect_text(value: &Value) -> String {
-    match value {
-        Value::Text(t) => t.to_string(),
-        other => panic!("Expected text, got {other:?}"),
-    }
-}
+mod common;
+use common::{expect_text, get_global, run_wfl};
 
 /// Create a temp file with some content and return its WFL-safe path string.
 fn temp_file(name: &str) -> (tempfile::TempDir, String) {
