+++ tests/http_stream_paced_test.rs
@@ -18,10 +18,8 @@ use std::time::Duration;
 use tokio::io::{AsyncReadExt, AsyncWriteExt};
 use tokio::net::TcpListener;
 
-use wfl::interpreter::Interpreter;
-use wfl::interpreter::value::Value;
-use wfl::lexer::lex_wfl_with_positions;
-use wfl::parser::Parser;
+mod common;
+use common::{get_number, get_text, run_wfl_ok as run_wfl};
 
 /// One chunked-encoding frame for `payload` + `\n`.
 fn chunk_frame(payload: &str) -> String {
@@ -31,42 +29,6 @@ fn chunk_frame(payload: &str) -> String {
 
 const CHUNKED_HEAD: &[u8] =
     b"HTTP/1.1 200 OK\r\nContent-Type: application/x-ndjson\r\nTransfer-Encoding: chunked\r\n\r\n";
-
-async fn run_wfl(code: &str) -> Interpreter {
-    let tokens = lex_wfl_with_positions(code);
-    let mut parser = Parser::new(&tokens);
-    let program = parser
-        .parse()
-        .unwrap_or_else(|e| panic!("Parse error: {e:?}"));
-    let mut interpreter = Interpreter::new();
-    interpreter
-        .interpret(&program)
-        .await
-        .unwrap_or_else(|e| panic!("Runtime error: {e:?}"));
-    interpreter
-}
-
-fn get_var(interpreter: &Interpreter, name: &str) -> Value {
-    interpreter
-        .global_env()
-        .borrow()
-        .get(name)
-        .unwrap_or_else(|| panic!("Variable '{name}' not found"))
-}
-
-fn get_number(interpreter: &Interpreter, name: &str) -> f64 {
-    match get_var(interpreter, name) {
-        Value::Number(n) => n,
-        other => panic!("Expected '{name}' to be a number, got {other:?}"),
-    }
-}
-
-fn get_text(interpreter: &Interpreter, name: &str) -> String {
-    match get_var(interpreter, name) {
-        Value::Text(t) => t.to_string(),
-        other => panic!("Expected '{name}' to be text, got {other:?}"),
-    }
-}
 
 /// A chunk that arrives strictly AFTER the program has consumed everything
 /// sent so far must wake the parked read and be delivered.
