+++ tests/password_hashing_test.rs
@@ -3,45 +3,8 @@
 // storing passwords. These builtins add slow, salted, memory/CPU-hard password
 // hashes that produce self-describing PHC/MCF strings and verify in constant time.
 
-use wfl::interpreter::Interpreter;
-use wfl::interpreter::value::Value;
-use wfl::lexer::lex_wfl_with_positions;
-use wfl::parser::Parser;
-
-/// Run WFL code and return the value stored in `result`.
-async fn run_wfl_code(code: &str) -> Result<Value, String> {
-    let tokens = lex_wfl_with_positions(code);
-    let mut parser = Parser::new(&tokens);
-    let ast = parser
-        .parse()
-        .map_err(|e| format!("Parse error: {:?}", e))?;
-
-    let mut interpreter = Interpreter::new();
-    interpreter
-        .interpret(&ast)
-        .await
-        .map_err(|e| format!("Runtime error: {:?}", e))?;
-
-    if let Some(result_value) = interpreter.global_env().borrow().get("result") {
-        Ok(result_value)
-    } else {
-        Err("Variable 'result' not found after execution".to_string())
-    }
-}
-
-fn expect_text(result: Result<Value, String>) -> String {
-    match result {
-        Ok(Value::Text(t)) => t.to_string(),
-        other => panic!("Expected text result, got {:?}", other),
-    }
-}
-
-fn expect_bool(result: Result<Value, String>) -> bool {
-    match result {
-        Ok(Value::Bool(b)) => b,
-        other => panic!("Expected bool result, got {:?}", other),
-    }
-}
+mod common;
+use common::{expect_bool_result as expect_bool, expect_text_result as expect_text, run_wfl_code};
 
 // === Generic password hashing (Argon2id default) ===
